Bug #29623286 Add AsyncLogger functionality [3/7]
The Logger mechanism is extended with the ability to invoke
all bound log handlers asynchronously. This can be activated
and deactivated at runtime as required.

This is implemented by passing log() calls to a buffered log
handler, which then invokes the LogListHandler in another thread.

The effect is to decouple the thread(s) generating log messages
from the thread consuming them and invoking the log handlers,
so that logging does not endanger thread liveness.

Async Logger functionality can be activated/deactivated through
the functions startAsync() /stopAsync() of logger instance.

Change-Id: I9be45f0837fde95dc0547f9fde3f56c8b772fc2d