Bug#35654240: A Recursive CTE Causing MySQL Server Crash
Problem: Right-nested UNIONs are not supported for recursive query
blocks in CTEs. A check for this is missing. In WL#11350, we started
allowing right-nested set operations. The missing check resulted in
a crash.

Solution: If there are right-nested UNIONs of recursive query blocks
that cannot be safely flattened, we throw an error indicating that
the operation is not yet supported.

Change-Id: I8ac01f91bf63e81fe7393179a3d01f9ba1eaa4e2