Bug#36027018 Testing function wait_connection_dropped always returns true.
wait_connection_dropped that is used in some component tests is
supposed to check if the connections via the Router are closed
as expected.

Because of the typo in the query in this function it was always
returning true even if the connection was not closed as expected.

Fixing the function revealed some other tests issues that this
patch is also addressing.

Change-Id: I9a730629a8e344fe8788fcee0071c85110b6d079