Bug#35970620 hash_set_operations optimizer off assertion error
Disabling hashed set operations after PREPARE caused EXECUTE to fail.
In MaterializeIterator<Profiler>::Init(), we have this code which
tries to open an index on the materialized output table of the set
operation:

if (doing_hash_deduplication() && !m_use_hash_map) {
    if (table()->file->ha_index_init(0, /*sorted=*/false)) {
      return true; <----
    }
}

This returned true without raising any error, hence the assert. The
reason for the error is that the table was prepared for hashed set
operation (the default), but has meanwhile been disabled
(m_use_hash_map reflects the user variable hash_set_operations).
This means TABLE::s->keys is zero, hence we can't open an index as
attempted. The solution is to close and re-open the tmp table with a
key and index.

Change-Id: Iaea06ae4c17cbafdb46e1030c97e551339c30ecd