Bug#36028618 add_trx_relevant_locks_to_cache wastes CPU and memory
The information_schema no longer has innodb_locks and innodb_lock_waits
tables for which the original `i_s_table_cache_t innodb_locks` was
populated. As of today the only lock related information is a single
information_schema.innodb_trx.trx_requested_lock_id column which
contains the id of the unique lock object for which this transaction is
waiting if any. There's no point in caching the information about
the lock requests of other transactions which are blocking the currently
processed one - we no longer report them anywhere in information_schema.

This patch removes the add_trx_relevant_locks_to_cache, and instead
directly adds the single lock the transaction is waiting for (if any) to
the innodb_locks cache when populating the row for that transaction.

The only reason this patch doesn't remove the innodb_cache as well, is
because without the cache the lock information would have to be stored
inline in the i_s_trx_row_t row, making it size pessimistically large.
With cache, we are using 8 bytes for pointer, and optionally 40 bytes
for the wait_lock description.
Without the cache, we would always use 48 bytes for
std::optional<i_s_locks_row_t> no matter if the trx is waiting or not.
In extreme case, removing the cache could increase space consumption by
20%, while the goal of this patch is to decrease it.

Change-Id: Idf5fe25eb96743d48cf1c4fc9edc083ba8ac951f