Bug#36028782 TRPMAN should only operate on Transporters in its own recv-thread
TRPMAN is intended to only operate on Transporters assigned
to its own recv-thread.

This requirement is broken when multiple transporters are set
up between the data nodes. In such cases the TRPMAN assigned to
operate on the first transporter get the responsibility for
all the others as well.

The problem roots back to when there used to be only a single
transporter between the data nodes. Then the nodeId could be used
just as well to identify the transporter / TRPMAN to be used.

Patch replace the Trpman::handles_this_node() with the method
::handles_this_trp().

TRPMAN Methods which may need to handle multiple transporters
between the same data nodes are enhanced to loop over the TrpId's
and check handles_this_trp() for each of them.

Some methods, as setting up the the initial connection, will
start to connect the single initial base transporter. These
methods will use the new method ::get_the_only_base_trp(NodeId)
which will look up the TrpId for the specified NodeId. It
will also 'require' that only a single TrpId is found.

Change-Id: I9b28a878fc36c50274920f41056929da50aab3f4