Bug#36028782 TRPMAN should only operate on Transporters in its own recv-thread
Remove usage of the TransporterRegistry methods which connect/disconnect
all the (Multi-)transporters to a NodeId, use the single TrpId variants
of these methods instead.

Note that this particular sub-patch does not fix a missuse of these
methods: All usage replaced in this patch had a 1::1 relation
between NodeId's and a single TrpId. Verified by using the method
get_the_only_base_trp(NodeId) to find the relevant TrpId to operate on.

However it is good to eliminate these methods potentially operating
on multiple node transporters which could lead to future missuse.

Change-Id: I138816d075d8408f4b3e8588fbc14ac7a7a0b1f5