Bug#36031560 Extend ndbinfo.transporters to show each transporter
Patch introduce the virtual ndbinfo.transporter_details table,
offering a per-transporter info of the communication links.

Information such as bytes_send and bytes_received is now offered
for each transporter.

Several TransporterRegistry get'ters are enhanced to take a TrpId
argument instead of the NodeId. These get'ters were used by the
existing ndbinfo.transporters table when aggregating information
over all the transporters between the same nodes.

Trpman::::execDBINFO_SCANREQ() now does this aggregation
instead for the TRANSPORTERS_TABLEID over all the transporters
for the node it will present a ndbinfo row for. For the
TRANSPORTER_DETAILS_TABLEID the enhanced get'ters are used directly
for each TrpId-row in the table.

Note that the overload and slowdown info is still maintained pr
node -> A transporter is regarded as overloaded if any of the
transporters to the same node is overloaded!
Fixing that will be a seperate patch.

Change-Id: I62dcc2a57f2af598c3cbad44726e199621ac6cc5
(cherry picked from commit ba3e52f96494d76b655af9266ee146acc2bdf022)