BUG#36012509: Add gr_gcs_message_data_psi_memory_spm test to 8.0
Group Replication (GR) rely on the Group Communication Service (GCS)
to exchange messages with the group and be informed of the group
membership.
When there is event from GCS, for example a message or membership
update, the GCS delivery thread will call the respective handler on
GR so that GR pursuits the required actions. The GCS delivery thread
is not initialized with a MySQL context. Some operations done on the
GR event handlers do require a MySQL context to be performed on
debug builds, the context is required for the debug information
provided by the performance schema instrumentation framework.
The operations that do require a MySQL context on debug builds are:
 * GR member actions load and store;
 * asynchronous replication channels failover configuration load and
   store;
 * debug flags for conditional execution.
To allow these operations, a temporary context was created, that is,
they were surrounded by:
```
  my_thread_init();
  ...
  my_thread_end();
```
These operations only happen on single-primary mode.

On 8.X.0 it was noticed that this temporary MySQL context was
causing side-effects, more precisely it was disabling the memory
consumption tracking inside GCS. This was fixed on 8.X.0.

On 8.0 the temporary MySQL context does not cause side effects.
Though this showed that this scenario was not being tested, thence
we are adding it so that it runs on regression testing.

Change-Id: Id042d4f207a88195f45d9e4ac87dbc2a5009672a