Bug#35957627: Mysql server fails for a particular query
While pushing condition down to a derived table with set operation,
currently mysql checks if the result types of columns present
in the where condition matches across all query blocks of the
set operation. If not it does not push the condition down.
However, it does not check if the result type of the column
in the outer query block matches with the one in the inner query
blocks. This leads to problems later, as the condition that is
pushed down inherits the result type of the columns from the cloned
columns (derived table columns), but the result type of the
expression is the same as the original expression.

Solution is not to push down when result types do not match.

Change-Id: Id9bf0f6fbc3704cee1437cdbccd597396351babe