Bug#35115601 InnoDB - Page [...] still fixed or dirty crash during shutdown
Post-push test fix
Increased size of the table to ensure second call to insert_direct. The
second call to insert_direct is required to test the error path in the
DDL.

Background:
Builder::insert_direct calls m_btr_load->latch and m_btr_load->release.
The call to release will buffer fix a page. The corresponding call to
latch will unfix the above page provided m_n_recs is non-zero. This
means that the first call to latch will not do anything.

In the first call to insert_direct, latch will do nothing, but release
will buffer fix the page. In the second call to insert_direct, the test
will simulate an error in the online log and force the DDL to error out.
This way the test checks the cleanup code in the DDL. In the bug, this
cleanup had missed the latch, causing the page to remain buffer fixed -
which was caught at shutdown.

Change-Id: I70db8539c42bb4c1761b56cca5dadf814b2c5020