Bug#29623286 NDB : NO LOGGING OF API DETECTED HEARTBEAT ISSUES [6/7]
Context:
Api nodes periodically send API_REGREQ signals to each of the data
nodes to which they are connected.
The Data nodes respond with API_REGCONF signals which the API uses
separately to determine the health of each data node.

Data nodes will log in the cluster log when any node is detected as
sending nothing for 2 or more consecutive heartbeat periods and,
when any node is detected as sending nothing for more than 4
consecutive heartbeat periods data nodes will log in the cluster
log and disconnect those nodes.

Problem:
Unlike datanodes API nodes do not log anything when a data node is
detected as sending nothing for any number of periods, or even when
it is disconnected as a result.

Solution:
Modify NdbApi ClusterMgr to log API/MGMD side when a connection
issue is detected.

Change-Id: Ib899f32c50971d52e48768391352834fec67f1fe