Bug#36041889 routertest restart_sharing fails sporadically
The router integration test "routing_sharing_restart" fails
sporadically in the "kill_...and_select".

The test-setup doesn't wait for the connections to be fully
established and placed into the connection pool before creating
a new connection to the backend.

On slow setup, this may lead to new connections being created
instead of taken from the pool which then breaks assumptions
the test makes later.

Change
======

- wait for the connection to be in the pool before opening
  a connection to the same backend

Change-Id: I478e548835bdadfcbf6fc03b148cab66041ac348