Bug#35846847: Assertion Failure in               /mysql-8.0.34/sql/sql_derived.cc:524
An assertion that checked if the referenced items
in an Item_ref object have consistent outer
reference information, failed when the reference
was of type OUTER_REF. For objects of type
Item_outer_ref, "depended_from" information is
set for the Item_outer_ref object and the
original expression that this reference points to.
However, there is an intermediate reference object
between Item_outer_ref and the original expression
which does not have this information.

Solution is to skip the checks for the Item_ref
object within the Item_outer_ref object.

Change-Id: Ide2fc4938b990cdd4636aee481617717dd6bcc99