Bug#35947358 Mysqld crash-my_valid_mbcharlen_utf8mb4
We have a rollup query with a window function (COUNT) in the select
list which we also order by. The wf is referenced by its alias in the
ORDER BY.  During the resolve phase, the function `split_sum_func' is
first run on the window function when we process the select list
(setup_fields). This which proceeds normally: the Item_field is
replaced with an `Item_aggregate_ref'. When resolving rollup
(resolve_rollup), COUNT's argument (now: the aggregate reference) is
wrapped in an Item_rollup_group_item.  Later during prepare, when we
get to the ORDER BY, we again run split_sum_func on COUNT. This action
was introduced by the patch listed in this bug issue as then regression
culprit.  Trying to split Item_rollup_group_item causes the problem,
placing it in the base_ref_items where it is not expected, which
causes the crash.

Solution: do not split Item_rollup_group_item.  If we encounter
Item_rollup_group_item during split_sum_func, we can be sure (I think)
split_sum_func has been run for its parent item already.

In a variant of the query, where the window function is *not* present
in the select list, there is no issue, because there, during ORDER
BY's split_sum_func phase, the window functions argument is still an
Item_field, which gets replaced by an Item_aggregate_ref and
eventually with a Item_rollup_group_item, this time during
resolve_rollup_wfs.

Change-Id: Icfd016b05fe31cd5e2b9e144d46f77a20fd509f5