Bug#33800633 startChangeNeighbour problem
2'nd attempt to fix this problem, 1'th patch has been reverted.

Transporters needing send-attention from a (assist-)send-thread
are returned from the method get_trp(). That method will give
extra priority to the transporters to neighbour node by returning
such neighbour transporters out of the regular TrpId fifo-queue
order inbetween returning non-neighbour transporters from the
head of the fifo-queue.

This has been implemented using two different ways of inserting
the TrpId's into the queue of trps needing a send. (::insert_trp()),
depending on whether they are a neighbour transporter or not:

 - non-neighbours are inserted into a fifo-list of transporters.
 - neighbour transporters are *not* inserted into any list, they
   are just annotated as 'having data'.

get_trp() in its side will either use the list of trps to return
a non-neighbour, or scan its set of neighbour transporters in
order to find a trp having available data to send, and not yet
assigned to another send-thread. Note that in the later case
the TrpId is not in the trp-fifo-list.

It is a prereq. in the send-thread protection mechanism that the same
TrpId should not be operated upon by two different send-threads.
Thus, when get_trp() has returned a TrpId it should not be get'ted
by another send thread until the first thread has completed its
send and the TrpId has been inserted into the queues again with
insert_trp(). The above also requires that a neighbour transporter
should never be in the TrpIs-fifo-queue: That would allow it to
be returned both as a neigbour and as a non-neighbour to two
different send threads.

During connect/disconnect of multi transporters we use the
'ChangeNeighbour' methods to set up new neighbours for the
(dis)connecting transporters. At this point we may need to
make sure that TrpIds waiting in the the two different queue
mechanism switch over to use the other type of queue.

In its original implementation we did not attempt to handle
this at all. Thus the bug report for Bug#33800633 claimed that
send to transporters left out could stall.

In the patch now being reverted we used insert_trp() to insert
old neighbour transporters not being a neighbour any more, into
the fifo-queue of non neighbours in startChangeNeighbour(). However
this opened up for a double representation of this TrpId if
setNeighbourNode() later added the same node back as a neighbour.

-> Having two different ways of representing neighbour vs non-neighbour
TrpId's in the queue system is a bug magnet!

This patch redesign this queueing mechanism such that all TrpId's
needing a send is now put into a TrpId-list.

The representation of neighbour transporters is kept mostly
unchanged. However it now works on top of the TrpId list, where
get_trp() now trying to find a 'is_enqueued(TrpId)' among the neighbour
when searcing for a high-priority TrpId to return.

The list implementation itself is changed to a double linked list
as we now need to remove neighbour TrpId's from the middle of the list.

thr_send_trps::m_thr_no_sender member, only used to check that the TrpId
was not yet assigned to any send-thread when get'ing a neighbour
transporter has been removed. It is replaced by checking that the
TrpId 'is_enqueued'. The asserts using m_thr_no_sender to check that
TrpId is not taken by other threads are changed the same way.

Updated (removed) comments describing how we used to 'lock'
the neighbour transporters in special ways.

No change should be required to the 'ChangeNeighbour' methods as the
TrpId's should still be in the same queue, independent of its neighbour
state. We should just rebuild the list of neighbours selecting from
this queue, as we already do.

Change-Id: I6429623dcb6c93c5ff59ba4dee01c0a53add9755