WL#16085 Record usage of information_schema.processlist
To get an impression of the usage of i_s.processlist,
we add status counters to the mysqld server to keep
track of how many times i_s processlist is used in a
query, and the last timestamp for when it was used.

We also add a thread in the health monitor component
to print the variable values to the error log once an
hour, provided the usage is > 0.

This makes us able to use log searching in MHS to see
how frequently i_s processlist is used.

Change-Id: Ia1060151cd265fd81d37e88c5c9d3d30c703bae9