Bug#36039546 main.histograms failing on pb2 daily-trunk
WL#15786 Automatically updated histograms introduced a call to
c_ptr_safe() on String objects (representing column names) allocated on
a MEM_ROOT by the parser. This resulted in additional heap allocations
without matching calls to free(), causing a memory leak detected by
valgrind.

This patch removes the problematic call to c_ptr_safe() and instead
copies the contents of the String object to a null-terminated C-style
string that is allocated on thd->mem_root.

This patch also fixes another (minor) problem introduced by WL#15786:
The MTR test histogram_auto_update depends on the EXPLAIN JSON format of
the optimizer when verifying that updated histograms are being used.
Since the test was written for the original optimizer's EXPLAIN JSON
format it failed when running MTR with the hypergraph optimizer enabled.
We change the test to explicitly use the hypergraph optimizer instead.
The infrastructure for histograms is independent of the choice of
optimizer so it suffices to test automatically updated histograms under
one of them.

Change-Id: I8072c48096f07414bb306c49941d27ee2c07ede5