Bug#34155137 STRAIGHT_JOIN gives different order for hash join with hypergraph
The hypergraph optimizer constructs hash joins by using the left-
side table as the probe table and the right-side table as the build
table. This is incorrect for straight joins, since straight joins
require that the left-side table should always be read before the
right side table. As the build table is read first, the left-side
table needs to be used as the build table for straight joins.

This is fixed by adding a check for straight joins before proposing
a hash join. If the join is a straight join, the left-side table is
proposed as the build table and the right-side table as the probe
table.

Change-Id: I7b37969c1494b842502624a395ff1ac48d41215d