Bug#35991881 Hypergraph: Wrong init_cost for nested loop join
This commit ensures that the init_cost of a nested loop join includes
the init_costs of both the inner and outer subpaths. Prior to this
fix, only the init_cost of the outer subpath was used. This caused
some NLJ paths to get an unreasonably low init_cost.

init_cost should be the initialization cost prior to reading the first
row, so that the cost to read k out of N rows would be
init_cost + (k/N) * (cost - init_cost). And for a NLJ, you cannot produce
the first result row without both reading the first outer row and trying
to read the first inner row (given that there was an outer row to be read).

When testing with TPC-DS (scale factor 1), fixing this cuts running
time of query 31 by 40%. The running time of query 69 is cut from > 5
minutes (the test timeout) to ~1.3s.

Change-Id: Ie57fee1889509052ebc94b9a6ce09630af57a237