Bug#36016140 Contribution: On s390x, compile the FMA test with -O2 to avoid over-optimiz ...
This patch is a contribution from:
Jonathan Albrecht <jonathan.albrecht@ibm.com>

Note:
s390x is not an officially supported platform for MySQL.
The patch is applied verbatim, without any testing.

Commit message from contributor:
===============================
The test for fused multiply add is compiled with -O3 by default but
on s390x with g++ 11 or 12 at least, -O3 will optimize out the math
instructions so the FMA optimization doesn't get a chance to be
applied. Reduce the optimization level to -O2 so FMA can be detected.

This allows some of the mtr tests to pass that otherwise fail if FMA
is applied.

Change-Id: I245bff6b2b898a778b4b827c3f6096e44c172765
(cherry picked from commit 8cedab567b6547aa82f88fe8287121d4b4801c11)