Bug#36053241 Background histogram update releases MDL early
WL#15786 Automatically updated histograms introduced a bug where MDL is
released on dictionary objects before they are removed from the
dictionary cache. This triggered assertions in some rapid test cases on
cloud builds.

This patch ensures that we first release objects from the dictionary
cache and then release MDL when updating histograms in the background.
The fix swaps the construction order of the scope guard objects that
handle these tasks to ensure we get the order right upon exiting the
function that handles background histogram updates.

Change-Id: I4583ae871d0029835125dba5ab2bfd21cedbbaab