BUG#36054134 Failure to run start_hook error
Problem:
Starting replica to apply changes when NDB is not ready or started the
reported error message is not helpful:

  mysql> SHOW REPLICA STATUS
  ...
  Last_errno: 13117
  Last_error: Fatal error: Failed to run 'applier_start' hook

Analysis:
When replica starts, the applier start hook first wait for NDB to become
ready for the number of seconds specified by --ndb-wait-setup= has
expired. If NDB is not ready by then, the start hook returns a failure
and the unhelpful error message is returned.

Solution:
Don't return an error from start hook when NDB is not ready, instead let
processing continue and error be returned from applying changes to NDB.
Thus setting an error message which better indicates the cause of the
error.

Change-Id: Ifb376f5f6ba43ed42c762579b9fc44978e58eded