Bug#36039905: Assertion `table_num_to_node_num[table_num] != -1' failed.
Some queries that called GROUPING in a subquery hit an assertion in
the hypergraph optimizer when running as prepared statements.

The problem was that Item_func_grouping::update_used_tables() used
LEX::current_query_block() to calculate the used tables, but when it
was called by code that reset prepared statement properties,
current_query_block() returned the top-level query block instead of
the subquery's query block. Item_func_grouping::used_tables()
therefore returned a wrong value, and it was detected by an assertion
in the hypergraph optimizer.

Fixed by saving a pointer to the query block in Item_func_grouping
during resolving, and using the saved pointer instead of
LEX::current_query_block() when updating the used tables.

Change-Id: I764b5e7e2f9683a9276977265a2fb2cb68b55471