Bug#36039546 main.histograms failing on pb2 daily-trunk
Post-push fix.

The fix for bug#36039546 included a change of the histogram_auto_update
MTR test to use the hypergraph optimizer instead of the original
optimizer. This change introduced a different combination of statements
to trigger automatic histogram updates that turned out to be unstable.

This patch reverts the test back to the previous version of the test
that uses the original optimizer, but ensures that the test also runs
with the hypergraph optimizer by adding --skip_if_hypergraph on
statements that depend on the optimizers EXPLAIN format.

Change-Id: I7569264f1f7d8c7f9c6eecc25272461f7be5a027