Bug#35476172: MySQL server fails when executing query
The problem here is when a partition clause in a CREATE TABLE statement
contains a subquery.
This is not valid syntax, however our problem here is that the subquery
is attached to the main query expression of the SQL command.
If this CREATE TABLE statement also contains a source query expression
for initial population of the table, this query expression will be
resolved before the table is created. But the mentioned subquery is an
incomplete structure and the resolving may thus fail.

The fix for the problem is to disallow subqueries in partition
expressions by clearing the member LEX::expr_allows_subselect while
contextualizing the partition expression. This means that such invalid
expressions will cause a syntax error.

Also renamed the expr_allows_subselect member to the more standard
looking expr_allows_subquery.

Change-Id: I40796351e37561404d372a10cd8e960c18535512