Bug#36061199 - Component tests cause dozens of pop up windows on macos.
When the Router's component tests are run on MacOS they cause plenty of pop up
windows from the OS firewall with a message:  "Do you want the application
mysq_server_mock to accept the incoming connections?"
They disappear after a while but they are quite annoying when one
is using a OS while running the tests.

The reason for those pop ups is the processes launched by the test binary that
are binding to the 0.0.0.0 address to accept connections. The MacOS firewall
does not like it.

This patch changes the tests to bind using 127.0.0.1 most of the times.
After this changes there are still 2 tests left that are triggering a couple of
pop ups:

1) routertest_component_http_server
these are the tests that specifically are testing binding to 0.0.0.0

2) routertest_component_state_file
there is a few tests there that are binding to ipv6 :: address. Changing those
would require a bigger refactoring.

Change-Id: Id5bde3b2243a7d763194bf87f496c8735d7c1494