Bug#34155137: STRAIGHT_JOIN gives different order for hash join with hypergraph
Post-push fix.

main.with_recursive times out when running with the hypergraph
optimizer. The reason is that a forced join order produces an invalid
plan if a hash join is chosen, and the hypergraph doesn't reject such
hash joins. The resulting plan never terminates.

Fixed by teaching the hypergraph optimizer not to produce hash joins
with a recursive reference on the "build" side.

Change-Id: Ic159f112a19ed7eeed7c44862dc7af18fa66d9d1