Bug#36008133 Assertion `false' failed in  replace_embedded_rollup_references_with_tmp_fields
Regression cause by Bug#35390341 Assertion `m_count > 0 && m_count >
m_frame_null_count' failed.

That issue involved setting a user variable inside the argument of a
window function, which in turn was evaluated using the window frame
buffer (Window::m_needs_frame_buffering) row optimizable
(Window::m_row_optimizable). The first solution had an unfortunate
side-effect as seen by the present bug: here, the window function
doesn't have an argument containing the setting of a user variable, it
is the other way around: the setting of a user variable requires the
value of a window function.

Solution: refine the criterion for when we evaluate the setting of
user variable before windowing to only include the case of a wf
containing a setting (Bug#35390341).

Change-Id: Idc6824adf4bd123775a14b92bfe54824acf105c8