Bug#36060557 WL#15540: Expression containing Non-deterministic function UUID() is transformed
A scalar subquery containing UUID gets transformed to join with a
derived table if transform is enabled - this is not correct according
to the specifications for WL#15540.

For non-correlated queries, we also do not (and should not) transform
a subquery containing a non-deterministic function; actually the repro
query here isn't correlated, cf. WL#12885 requirement FR#9.

Solution: mark as uncacheable using the same approach as RAND.
Change-Id: Ia06417bf662e8687c63e5a9b69aadc7a02054c60