Bug#36075756 ASAN crash on MaterializeIterator<Profiler>::load_HF_row_into_hash_map() [1/2 noclose]
To see this error, we needed to run with ASAN compiled in and
--sanitize on the repro test case. It didn't trigger an ASAN error as
such, but hit an assert. Possibly this happens only with ASAN because
space usage increases on the heap due to ASAN padding before and after
allocated objects.

The error is another instance of the third issue solved in Bug#35686098
Assertion `n < size()' failed in Element_type& Mem_root_array_YY:

(quote):
  "we ran out of space in the dedicated hash table upon re-reading one
   of the on-disk chunks into the hash table."

The fix there was to use the general mem_root instead of the dedicated
one for the hash set operation for that particular corner case.
Specifically, we used the general mem_root when allocating the hash
map *payload*.

In this bug, we see the same issue: we can't fit the last row (#10 out
of 10 long varchar strings, i.e. blobs in the repro) into the hash
table mem_root: this time we run out when we allocate space for the
hash map *key*, in check_unique_fields_hash_map prior to calling
ImmutableStringWithLength::Encode(<hash key>).  The solution is the
same, use the general mem_root for this case: i.e. we generalize the
solution from Bug#35686098 to cover running out of space for both the
key and payload.

Change-Id: Id869e6afad7c996950779c008dcf650e21b17027