Bug#36061197 SQL thread fails with error MY-001755 when executed with parallel applier
Problem
-------

When a user copies binary logs from one server to the location
of the relay logs on another server, using direct file-copy instead
of receiving the files over a replication connection, the applier
stops with an error while processing the first transaction in the
first binary log produced after a restart of the source server.
The error message is: "Cannot execute the current event group in the
parallel mode. Encountered event Gtid, relay-log name
./binary-log.000069, position 197 which prevents execution of this
event group in parallel mode. Reason: The source event is logically
timestamped incorrectly".

Background
-------

The coordinator tracks the sequence_number of the last scheduled
transaction, and uses it to verify that sequence_number increases
monotonically.

The sequence_number always starts at 1 in the beginning of each
binary log on the source server, and then increments by 1 for each
transaction within the file. Therefore, the coordinator skips the
check for each transaction that occurs after a binlog file boundary
on the source.

The coordinator used Rotate events originating from the source to
detect file boundaries. This works for relay logs, because the
source always sends a rotate event. Even if a Rotate event is
missing in the binary log, which may happen e.g. if the source
server was restarted, the source generates a "fake" Rotate event.

Analysis / Root-cause analysis
------------------------------

When binary logs have been copied from a source server and used
by the applier instead of replicated relay logs, there are no
"fake Rotate" events. Therefore, in case the source that generated
the binary log was restarted while writing the binary log, so that
there is no Rotate, the applier does not suppress the check for
monotonic sequence_numbers. So when it processes the first
transaction in the new binary log, it conducts the check, which
fails, and therefore it stops with this error.

Solution
--------

Suppress the check for monotonic sequence_number when processing
the first Format_description_log_event of a binary log originating
from the source, rather than when processing a Rotate event. There
is always a Format_description_log_event.

Change-Id: I9d0200f914a77dc0a4ac493a2f82118c68ff76bf