BUG#36076308: Missing exit_state_action handling in two error situations
It was found that two error scenarios that would cause a member to
exit a group and move to ERROR state were not honoring the option
group_replication_exit_state_action:
 1) error while enabling super_read_only during member join;
 2) member join recovery impossible due to missing binary logs and
    clone plugin on group members.
On these scenarios the member changed to ERROR state and left the
group, but did not perform the group_replication_exit_state_action.
For instance, if group_replication_exit_state_action=OFFLINE_MODE,
offline_mode was not enabled.

To solve the above issues, the group_replication_exit_state_action
is now performed.

Change-Id: Ieda4cce71ac5a56104c106ef0b9b094732836a79