Bug#36070542 unnest scalar subquery using groupby produces wrong result
The issue shows that a transform can be semantically wrong with the current approach
if the SELECTed item in the subquery tests for NULL: the left outer join
with a grouped derived table might in such cases produce NULLs whereas the original
subquery might not. E.g.

   SELECT (SELECT IFNULL(SUM(t3.b), 0) FROM t3 WHERE t1.a=t3.a) FROM t1;

If the inner where clause were false, the subquery would produce 0 for
such rows in t1. The transformed subquery with our current approach
would be:

   SELECT `derived_1_2`.`IFNULL(SUM(t3.b), 0)` AS `subq`
   FROM `mysql`.`t1`
        LEFT JOIN
        ( SELECT IFNULL(sum(`mysql`.`t3`.`b`),0) AS `IFNULL(SUM(t3.b), 0)`,
                `mysql`.`t3`.`a` AS `a`
          FROM `mysql`.`t3`
          GROUP BY `mysql`.`t3`.`a`) `derived_1_2`
        ON `derived_1_2`.`a` = `mysql`.`t1`.`a`

This would produce NULLs for every row in t1 of the join predicate is
false.

Solution: bar such subqueries from being transformed
Change-Id: I459279a9cb66af918f6f0fc76288cb7304843c83