WL#15989  Innodb: Allow clone across patch releases within one LTS Series
Description:
  This worklog relaxes clone version constraints and allows cloning in
certain special cases when donor and recipient versions differ. Cloning
is allowed if the version strings match. For 8.1+ version series,
cloning is allowed if major and minor versions match. For 8.0 version
series, cloning is not allowed if either donor or recipient's patch
version is less than backport version (the version in which this worklog
is backported) and they mismatch. For 8.0 version series, if both patch
versions are above backport, then cloning is allowed if major and minor
versions match.

Change-Id: Ife21d4ab9d81434e296c5e856e56c75cc67f486e