Bug#36094867: Intersect decimal precision assert error
The problem is with the function my_decimal_length_to_precision()
that interprets a decimal scale of 31 (DECIMAL_NOT_SPECIFIED)
as a positive scale. However, the scale of a YEAR value should
always be zero, so this will be reflected in Item::set_data_type_year().

Also setting decimals to zero for all integer types, to ensure the scale
is set properly for these types too.

Change-Id: I17615780bed98fa405f8768d94a29a7a14cb448a